requirements.txt
# Core dependencies
web3==6.15.1
python-dotenv==1.0.0
requests==2.31.0

# Polymarket SDK
py-clob-client==0.25.0

# Data & Analysis
pandas==2.1.4
numpy==1.26.3

# Async support
aiohttp==3.9.1

# Logging & Monitoring
colorlog==6.8.0
rich==13.7.0

# UI Dashboard
streamlit==1.31.0
plotly==5.18.0

# AI & Analysis (optional - uncomment if needed)
# openai==1.6.1
# anthropic==0.7.8

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utils
python-dateutil==2.8.2
pytz==2023.3
